--- src/lib/ui/AdminDeployment/index.ts
@@ -1,3 +1,0 @@
-export { default as AdminDeploymentList } from './AdminDeploymentList.svelte';
-export { default as AdminDeploymentDetail } from './AdminDeploymentDetail.svelte';
-export { default as AdminDeploymentCreateLayer } from './AdminDeploymentCreateLayer.svelte';
