+++ src/lib/ui/AdminStructure/ImportNotebookMembers.svelte
@@ -12,9 +12,8 @@
 <script lang="ts">
 	import Dropzone from 'svelte-file-dropzone';
 	import { v4 as uuidv4 } from 'uuid';
-	import Checkbox from '$lib/ui/base/GroupCheckbox.svelte';
 	import { Text } from '$lib/ui/utils';
-	import { Alert, Button } from '$lib/ui/base';
+	import { Alert, Button, GroupCheckbox as Checkbox } from '$lib/ui/base';
 	import { page } from '$app/stores';
 	import { parse as csvParse } from 'csv-parse/browser/esm/sync';
 	import { pluralize } from '$lib/helpers';
