+++ src/lib/ui/Deployment/index.ts
@@ -1,0 +1,5 @@
+export { default as AdminDeploymentList } from './List.svelte';
+export { default as AdminDeploymentDetail } from './Detail.svelte';
+export { default as AdminDeploymentCreateLayer } from './CreateLayer.svelte';
+export { default as AdminDeploymentNotebookUpdate } from './NotebookUpdate.svelte';
+export { default as AdminDeploymentNotebookListUpdater } from './NotebookListUpdater.svelte';
