+++ src/lib/ui/Manager/View.svelte
