+++ src/lib/ui/Manager/_importStructure.gql
