+++ src/routes/admin/index.svelte
@@ -3,9 +3,8 @@
 	import type { Segment } from '$lib/routes';
 
 	import { openComponent } from '$lib/stores';
-	import { AdminDeploymentList, AdminDeploymentCreateLayer } from '$lib/ui/AdminDeployment';
-	import { Button } from '$lib/ui/base';
-	import Breadcrumbs from '$lib/ui/base/Breadcrumbs.svelte';
+	import { AdminDeploymentList, AdminDeploymentCreateLayer } from '$lib/ui/Deployment';
+	import { Breadcrumbs, Button } from '$lib/ui/base';
 	import { operationStore, query } from '@urql/svelte';
 
 	const deploymentsStore = operationStore(GetDeploymentsDocument);
